refactor(bridge): split oversized bridge implementation by feature
